agents/requirements.txt
# Claude AI Integration
anthropic==0.40.0

# Agentcy Framework
langgraph==0.4.1
langchain==0.3.0
langchain-anthropic==0.3.0

# Agent Communication
slim-transport==0.6.1
nats-py==2.11.8

# Model Context Protocol (MCP)
mcp==1.10.0

# Observability
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0

# HTTP Client
aiohttp==3.9.0
httpx==0.27.0

# Data Processing
numpy==1.24.0
pydantic==2.7.0

# Database
pymongo==4.7.0
redis==5.0.0

# Vector Database
qdrant-client==1.10.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
